Move command handlers to submodule
